requirements.txt
google-adk
google-genai>=0.3.0
tornado>=6.4
streamlit>=1.32.0
python-dotenv>=1.0.0
pydantic>=2.7.0
pydub>=0.25.1
PyPDF2>=3.0.0
google-generativeai>=0.7.0

# Authentication dependencies
bcrypt>=4.1.2
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
email-validator>=2.1.0